pyproject.toml
[project]
name = "agent-experiments"
version = "0.1.0"
description = "A collection of experiments with agents using smolagents"
readme = "README.md"
requires-python = ">=3.11"
authors = [
{ name = "Martin Hatting Petersen", email = "petersen2630@gmail.com" },
]
dependencies = [
    "smolagents[litellm,mlx-lm]>=1.5.0",
    "markdownify>=0.14.1",
    "duckduckgo-search>=7.2.1",
    "transformers>=4.48.1",
    "ipywidgets>=8.1.5",
    "python-dotenv>=1.0.1",
    "selenium>=4.28.1",
    "helium>=5.1.0",
    "pillow>=11.1.0",
    "arize-phoenix>=7.12.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "openinference-instrumentation-smolagents>=0.1.4",
    "matplotlib>=3.10.1",
    "geopandas>=1.0.1",
    "shapely>=2.0.7",
    "plotly>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipython>=8.31.0",
    "ipykernel>=6.29.5",
]
